Image is building; WIP

+++ jikesrvm-hg/MMTk/src/org/mmtk/utility/Log.java
@@ -146,20 +146,16 @@ public class Log implements Constants {
     int index = TEMP_BUFFER_SIZE - 1;
     char[] intBuffer = getIntBuffer();
 
-    // fixme: remove inner int cast; need 64bit modulus
-    nextDigit = (int) ((int)l % 10);
+    nextDigit = (int) (l % 10);
     nextChar = hexDigitCharacter[negative ? - nextDigit : nextDigit];
     intBuffer[index--] = nextChar;
-    // fixme: remove inner int cast; need 64bit divide
-    l = (int)l / 10;
+    l = l / 10;
 
     while (l != 0) {
-        // fixme: remove inner int cast; need 64bit modulus
-      nextDigit = (int) ((int)l % 10);
+      nextDigit = (int) (l % 10);
       nextChar = hexDigitCharacter[negative ? - nextDigit : nextDigit];
       intBuffer[index--] = nextChar;
-      // fixme: remove inner int cast; need 64bit divide
-      l = (int)l / 10;
+      l = l / 10;
     }
 
     if (negative) {
